+++ src/onegov.election_day/onegov/election_day/formats/vote/wabstic.py
@@ -52,7 +52,7 @@ def line_is_relevant(line, domain, district, number):
     )
 
 
-def import_vote_wabstic(vote, principal, district, number,
+def import_vote_wabstic(vote, principal, number, district,
                         file_sg_geschaefte, mimetype_sg_geschaefte,
                         file_sg_gemeinden, mimetype_sg_gemeinden):
     """ Tries to import the files in the given folder.